+++ .storybook/welcome-page/selling-points/index.js
@@ -1,0 +1,1 @@
+export { default } from './selling-points.component';