+++ .storybook/welcome-page/selling-points/selling-points-item.component.js
@@ -1,0 +1,6 @@
+import React from 'react';
+import { SellingPointItem } from './selling-points.style';
+
+export default ({ children }) =>
+  <SellingPointItem>{ children }</SellingPointItem>
+;