+++ cypress/locators/flat-table/locators.js
@@ -1,0 +1,4 @@
+// component preview locators
+export const FLAT_TABLE_COMPONENT = '[data-component="flat-table"]';
+export const FLAT_TABLE_ROW = '[data-element="flat-table-row"]';
+export const FLAT_TABLE_STICKY_ROW_HEADER = '[data-element="flat-table-row-header"]';