+++ cypress/support/step-definitions/dialog-full-screen-steps.js
@@ -1,4 +1,5 @@
 import { dialogFullScreenChildren, dialogFullScreenPreview, dialogFullScreenPreviewClosedState } from '../../locators/dialog-full-screen';
+import { dialogStickyFormFooter } from '../../locators/dialog';
 
 Then('Dialog Full Screen children on preview is {string}', (children) => {
   dialogFullScreenChildren().should('have.text', children);
@@ -10,4 +11,9 @@ Then('Dialog Full Screen is visible', () => {
 
 Then('Dialog Full Screen is not visible', () => {
   dialogFullScreenPreviewClosedState().should('exist');
+});
+
+Then('Dialog Full Screen stickyFormFooter is visible', () => {
+  dialogStickyFormFooter().should('exist')
+    .and('be.visible');
 });