+++ src/__experimental__/components/form/form.spec.js
@@ -501,6 +501,22 @@ describe('Form', () => {
           TestUtils.Simulate.submit(form);
           expect(spy).not.toHaveBeenCalled();
         });
+
+        describe('with sticky footer enabled', () => {
+          it('regains focus on submit button', () => {
+            instance = TestUtils.renderIntoDocument(
+              <Form stickyFooter>
+                <Textbox
+                  validations={ [new Validation()] } name='test'
+                  value=''
+                />
+              </Form>
+            );
+            const form = TestUtils.findRenderedDOMComponentWithTag(instance, 'form');
+            TestUtils.Simulate.submit(form);
+            expect(document.activeElement).toBe(form.getElementsByTagName('button')[1]);
+          });
+        });
       });
     });
   });