+++ src/components/icon-button/index.d.ts
@@ -1,0 +1,2 @@
+export { default } from './icon-button';
+export * from './icon-button';