+++ src/components/message/message.style.js
@@ -2,6 +2,7 @@ import styled, { css } from 'styled-components';
 import PropTypes from 'prop-types';
 import BaseTheme from '../../style/themes/base';
 import messageClassicStyling from './message-classic.style';
+import StyledIconButton from '../icon-button/icon-button.style';
 
 const MessageStyle = styled.div`
   position: relative;
@@ -16,6 +17,16 @@ const MessageStyle = styled.div`
       border: none;
       background: transparent;
   `}
+
+  ${StyledIconButton} {
+    right: 20px;
+    top: 50%;
+    transform: translateY(-50%);
+
+    &::-moz-focus-inner {
+      border: none;
+    }
+  }
 
   ${messageClassicStyling}
 `;